MAJOR: Refactor slug and variants dynamic routes. Implement variants selector with preserved state
