+++ app/(main)/[slug]/[[...variant]]/page.tsx
@@ -35,9 +35,8 @@ export async function generateStaticParams() {
 
   const params = species.flatMap((specie) =>
     specie.varieties.map((variant) => ({
-      slug: variant.is_default
-        ? [specie.name]
-        : [specie.name, variant.pokemon.name],
+      slug: specie.name,
+      variant: variant.is_default ? undefined : [variant.pokemon.name],
     }))
   )
 
@@ -47,11 +46,11 @@ export async function generateStaticParams() {
 export async function generateMetadata({
   params,
 }: {
-  params: Promise<{ slug: string }>
+  params: Promise<{ slug: string; variant: string[] | undefined }>
 }): Promise<Metadata> {
-  const { slug } = await params
-  const [speciesKey, variantKey] = slug
-  const species = await pokeapi.getPokemonSpeciesByName(speciesKey)
+  const { slug, variant } = await params
+  const [variantKey] = variant ?? []
+  const species = await pokeapi.getPokemonSpeciesByName(slug)
   const pokemon = await pokeapi.getPokemonByName(
     species.varieties.find((v) =>
       variantKey ? v.pokemon.name === variantKey : v.is_default
@@ -106,11 +105,11 @@ export async function generateMetadata({
 export default async function Page({
   params,
 }: {
-  params: Promise<{ slug: string }>
+  params: Promise<{ slug: string; variant: string[] | undefined }>
 }) {
-  const { slug } = await params
-  const [speciesKey, variantKey] = slug
-  const species = await pokeapi.getPokemonSpeciesByName(speciesKey)
+  const { slug, variant } = await params
+  const [variantKey] = variant ?? []
+  const species = await pokeapi.getPokemonSpeciesByName(slug)
   const pokemon = await pokeapi.getPokemonByName(
     species.varieties.find((v) =>
       variantKey ? v.pokemon.name === variantKey : v.is_default
